+++ ehospital.client/src/services/api/apiDoctorService.jsx
@@ -17,10 +17,11 @@ export const getDoctors = async (page, size) => {
 };
 
 export const getDoctorById = async (doctorId) => {
+    if (!doctorId) return;
     try {
-        console.log("get", doctorId)
+        //console.log("get", doctorId)
         const response = await axios.get(apiConfig.DOCTORS_URL + "/" + doctorId)
-        console.log("now", response.data)
+        //console.log("now", response.data)
 
         return response.data;
     } catch (err) {
@@ -29,6 +30,7 @@ export const getDoctorById = async (doctorId) => {
 }
 
 export const getDoctorsBySpeciality = async (speciality) => {
+    if (!speciality) return;
     try {
         const response = await axios.get(apiConfig.DOCTORS_URL + "/speciality/" + speciality)
         return response.data;
@@ -38,6 +40,7 @@ export const getDoctorsBySpeciality = async (speciality) => {
 }
 
 export const createDoctor = async (doctorData) => {
+    if (!doctorData) return;
     try {
         const response = await axios.post(apiConfig.DOCTORS_URL, doctorData)
         toast.success("Doctor created successfully");
@@ -48,6 +51,7 @@ export const createDoctor = async (doctorData) => {
 }
 
 export const updateDoctor = async (doctorData) => {
+    if (!doctorData) return;
     try {
         const response = await axios.put(apiConfig.DOCTORS_URL + "/" + doctorData.id, doctorData)
         toast.success("Doctor updated successfully");
@@ -58,6 +62,7 @@ export const updateDoctor = async (doctorData) => {
 }
 
 export const deleteDoctor = async (doctorId) => {
+    if (!doctorId) return;
     try {
         const response = await axios.delete(apiConfig.DOCTORS_URL + "/" + doctorId)
         toast.success("Doctor deleted successfully");
