+++ ehospital.client/src/services/api/apiPatientService.jsx
@@ -16,6 +16,8 @@ export const getPatients = async (page, size) => {
 }
 
 export const getPatientById = async (patientId) => {
+    if (!patientId) return;
+    if (!patientId) return;
     try {
     const response = await axios.get(apiConfig.PATIENTS_URL + "/" + patientId)
     return response.data;
@@ -26,7 +28,8 @@ export const getPatientById = async (patientId) => {
 }
 
 export const createPatient = async (patientData) => {
-    console.log("patientData", patientData);
+    if (!patientData) return;
+    //console.log("patientData", patientData);
     try {
         const response = await axios.post(apiConfig.PATIENTS_URL, patientData)
         toast.success("Patient Account created successfully");
@@ -37,6 +40,7 @@ export const createPatient = async (patientData) => {
 }
 
 export const updatePatient = async (patientData) => {
+    if (!patientData) return;
     try {
         const response = await axios.put(apiConfig.PATIENTS_URL + "/" + patientData.id, patientData)
         toast.success("Patient updated successfully");
@@ -47,6 +51,7 @@ export const updatePatient = async (patientData) => {
 }
 
 export const deletePatient = async (patientId) => {
+    if (!patientId) return;
     try {
         const response = await axios.delete(apiConfig.PATIENTS_URL + "/" + patientId)
         toast.success("Patient Account deleted successfully");
